+++ src/main/java/com/example/cp_main_be/domain/member/user/domain/User.java
@@ -2,8 +2,8 @@ package com.example.cp_main_be.domain.member.user.domain;
 
 import com.example.cp_main_be.domain.avatar.avatar.domain.Avatar;
 import com.example.cp_main_be.domain.garden.garden.domain.Garden;
+import com.example.cp_main_be.domain.mission.diary.domain.Diary;
 import com.example.cp_main_be.domain.social.bookmark.domain.Bookmark;
-import com.example.cp_main_be.domain.social.diary.domain.Diary;
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 import jakarta.persistence.*;
 import java.time.LocalDateTime;
