+++ src/main/java/com/example/cp_main_be/domain/mission/diary/dto/request/CreateDiaryRequest.java
@@ -1,4 +1,4 @@
-package com.example.cp_main_be.domain.social.diary.dto.request;
+package com.example.cp_main_be.domain.mission.diary.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -8,9 +8,9 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
-@NoArgsConstructor
+@NoArgsConstructor // JSON 역직렬화를 위해 기본 생성자가 필요합니다.
 @AllArgsConstructor
-public class UpdateDiaryRequest {
+public class CreateDiaryRequest {
 
   @NotBlank(message = "제목은 필수 입력 항목입니다.")
   @Size(max = 100, message = "제목은 100자를 초과할 수 없습니다.")
@@ -19,8 +19,10 @@ public class UpdateDiaryRequest {
   @NotBlank(message = "내용은 필수 입력 항목입니다.")
   private String content;
 
+  // 이미지는 보통 S3 같은 곳에 먼저 업로드한 뒤, 그 URL을 받아와 저장합니다.
   private String imageUrl;
 
+  // 공개 여부는 선택사항으로, 값을 보내지 않으면 엔티티의 기본값(true)을 따릅니다.
   @NotNull(message = "공개 여부는 필수값입니다. (true/false)")
   private Boolean isPublic;
 }
