+++ src/main/java/com/example/cp_main_be/domain/mission/diary/dto/response/DiaryFeedItemResponse.java
@@ -1,6 +1,6 @@
-package com.example.cp_main_be.domain.social.diary.dto.response;
+package com.example.cp_main_be.domain.mission.diary.dto.response;
 
-import com.example.cp_main_be.domain.social.diary.domain.Diary;
+import com.example.cp_main_be.domain.mission.diary.domain.Diary;
 import com.example.cp_main_be.global.dto.AuthorResponse;
 import com.example.cp_main_be.global.dto.FeedItemResponse;
 import java.time.LocalDateTime;
@@ -8,15 +8,15 @@ import lombok.Getter;
 
 @Getter
 public class DiaryFeedItemResponse implements FeedItemResponse {
-  private Long postId;
   private final String postType = "DIARY";
-  private AuthorResponse author;
-  private String title;
-  private String content;
-  private String imageUrl;
-  private int likeCount;
+  private final Long postId;
+  private final AuthorResponse author;
+  private final String title;
+  private final String content;
+  private final String imageUrl;
+  private final int likeCount;
   private int commentCount; // Diary에 Comment 리스트가 있다고 가정
-  private LocalDateTime createdAt;
+  private final LocalDateTime createdAt;
 
   public DiaryFeedItemResponse(Diary diary) {
     this.postId = diary.getId();
