+++ src/main/java/com/example/cp_main_be/domain/mission/diaryimage/presentation/DiaryImageController.java
@@ -1,6 +1,6 @@
-package com.example.cp_main_be.domain.social.diaryimage.presentation;
+package com.example.cp_main_be.domain.mission.diaryimage.presentation;
 
-import com.example.cp_main_be.domain.social.diary.service.DiaryService;
+import com.example.cp_main_be.domain.mission.diary.service.DiaryService;
 import com.example.cp_main_be.global.common.ApiResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -12,7 +12,7 @@ import org.springframework.web.multipart.MultipartFile;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("api/v1/diaries")
+@RequestMapping("api/v1")
 @Tag(name = "일기 이미지 API", description = "일기의 이미지 관련 기능을 제공합니다")
 public class DiaryImageController {
 
@@ -31,7 +31,7 @@ public class DiaryImageController {
   }
 
   @Operation(summary = "일기 이미지 삭제", description = "일기 이미지를 삭제합니다")
-  @DeleteMapping("{diaryId}/images/{imageId}")
+  @DeleteMapping("/diaries/{diaryId}/images/{imageId}")
   public ResponseEntity<ApiResponse<Void>> deleteDiaryImage(
       @PathVariable Long diaryId, @PathVariable Long imageId) {
     diaryService.deleteDiaryImage(diaryId, imageId);
